#6 [refactor] : 프로젝트 구조 개선 및 공통 엔티티 구성요소 위치 변경
- BaseEntity를 domain.common 패키지로 이동
- Status를 domain.common.enums 패키지로 이동
- @Embeddable 클래스들을 domain.common.template 패키지로 이동
