+++ src/main/java/com/linku/backend/domain/icon/repository/IconRepository.java
@@ -1,7 +1,7 @@
 package com.linku.backend.domain.icon.repository;
 
 import com.linku.backend.domain.icon.Icon;
-import com.linku.backend.global.common.enums.Status;
+import com.linku.backend.domain.common.enums.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
