+++ src/main/java/com/linku/backend/domain/page/Page.java
@@ -1,6 +1,6 @@
 package com.linku.backend.domain.page;
 
-import com.linku.backend.domain.BaseEntity;
+import com.linku.backend.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
