+++ src/main/java/com/linku/backend/domain/subscribe/Subscribe.java
@@ -1,6 +1,6 @@
 package com.linku.backend.domain.subscribe;
 
-import com.linku.backend.domain.BaseEntity;
+import com.linku.backend.domain.common.BaseEntity;
 import com.linku.backend.domain.page.Page;
 import com.linku.backend.domain.user.User;
 import jakarta.persistence.*;
